feat: provide `--service-manager` argument
The `start` command now has a `--service-manager` flag argument, which if used will start the
services not using `antctl` but go directly via the service manager.

This is to support a specific testing scenario we have.

Also remove use of the `--force` argument in the upgrade playbook when a URL is being used. We can
just explicitly pass the argument, and it causes an error if it gets used twice.